app/src/main/java/com/lupesoft/pruebadeingreso/data/user/UserDao.kt
package com.lupesoft.pruebadeingreso.data.user

import androidx.lifecycle.LiveData
import androidx.room.Dao
import androidx.room.Insert
import androidx.room.OnConflictStrategy
import androidx.room.Query
import kotlinx.coroutines.flow.Flow

@Dao
interface UserDao {

    @Query("SELECT * FROM users ORDER BY name")
    fun getUsers(): Flow<List<UserEntity>>

    @Query("SELECT * FROM users ORDER BY name")
    fun getUsersTest(): LiveData<List<UserEntity>>

    @Query("SELECT * FROM users WHERE id = :userId")
    fun getUser(userId: Int): LiveData<UserEntity>

    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insertAll(userEntities: List<UserEntity>)

    @Insert(onConflict = OnConflictStrategy.IGNORE)
    suspend fun insertAll(vararg users: UserEntity): List<Long>

}